Add SimpleMap encoding type in serializer. Update test and sample. Add more API documentation for AmqpContract/Member/Provides usage.
